Implementation v1 of IgnoreNeed
Also ordered the classes into folders
