+++ Defs/Test.xml
@@ -1,0 +1,54 @@
+<?xml version="1.0" encoding="utf-8"?>
+<Defs>
+	<PawnKindDef ParentName="BasePlayerPawnKind">
+		<defName>TestColonist</defName>
+		<label>test	colonist</label>
+		<defaultFactionType>PlayerColony</defaultFactionType>
+		<chemicalAddictionChance>0.06</chemicalAddictionChance>
+		<apparelTags>
+			<li>IndustrialBasic</li>
+		</apparelTags>
+		<apparelAllowHeadgearChance>0</apparelAllowHeadgearChance>
+		<apparelMoney>350~600</apparelMoney>
+		<backstoryCryptosleepCommonality>1</backstoryCryptosleepCommonality>
+		<techHediffsChance>0.03</techHediffsChance>
+		<techHediffsMoney>50~800</techHediffsMoney>
+		<techHediffsTags>
+			<li>Poor</li>
+			<li>Simple</li>
+			<li>ImplantEmpireCommon</li>
+		</techHediffsTags>
+		<modExtensions>
+			<li Class="SPADE.DefExtension_DoesNotNeed">
+				<needs>
+					<li>Beauty</li>
+					<li>Comfort</li>
+					<li>Outdoors</li>
+					<li>RoomSize</li>
+				</needs>
+			</li>
+			<li Class="SPADE.DefExtension_NonStandardMedicine">
+				<medicine>Spade_Medicine</medicine>
+			</li>
+			<li Class="SPADE.DefExtension_CarriesWeaponOpenly"/>
+			<li Class="SPADE.DefExtension_CarriesWeaponStraight">
+				<offsets>
+					<north>(0, 0.036735, 0.11)</north>
+					<south>(0, 0, 0.11)</south>
+					<west>(0.2, 0, 0.33)</west>
+					<east>(-0.2, 0, 0.33)</east>
+				</offsets>
+			</li>
+			<li Class="SPADE.DefExtension_CarriesThingsOffset">
+				<offsets>
+					<north>(0, 0, 0.6)</north>
+					<south>(0, 0, -0.6)</south>
+					<west>(-0.6, 0, 0)</west>
+					<east>(0.6, 0, 0)</east>
+				</offsets>
+			</li>
+			<li Class="SPADE.DefExtension_SPADE_Resurrectable"/>
+			<li Class="SPADE.DefExtension_Repairable"/>
+		</modExtensions>
+	</PawnKindDef>
+</Defs>
