+++ Source/communityframework/communityframework/IgnoreNeed.cs
@@ -1,0 +1,15 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+using RimWorld;
+using Verse;
+
+namespace CF
+{
+    class IgnoreNeed : DefModExtension
+    {
+        public List<NeedDef> needs;
+    }
+}
