+++ Source/communityframework/communityframework/RemoteTrigger/CompProperties_RemoteTrigger.cs
